CleanCode-2 + code anh thựn

+++ src/main/java/com/VooTreeVeeVuu/domain/repository/RoomImageRepository.java
@@ -3,5 +3,5 @@ package com.VooTreeVeeVuu.domain.repository;
 import com.VooTreeVeeVuu.domain.entity.RoomImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoomImageRepository extends JpaRepository<RoomImage, Integer> {
+public interface RoomImageRepository extends JpaRepository<RoomImage, Long> {
 }
